Move KalmanFilter codes to separate files
